数据库支持

+++ app/src/main/java/com/qi/billiards/game/DePlayer.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 data class DePlayer(
     var name: String,
-    var score: Int, // 当前分数
+    var score: Double, // 当前分数
     var cost: Double = 0.0, // 台费
     var buyCount: Int = 0, // 买入次数
     var size: Int = 1, // AA台费
